// condor/client/ng/search/common/search-operators.ts
export enum SearchOperator {
    equalTo = '0',
    notEqualTo = '1',
    startsWith = '2',
    endsWith = '3',
    contains = '4',
    exists = '5',
    notExists = '6',
    between = '7',
    notBetween = '8',
    soundsLike = '9',
    lessThan = '10',
    lessEqual = '11',
    greater = '12',
    greaterEqual = '13',
    withinLast = '14',
    withinNext = '15',
    specificDate = 'sd'
}